Move setting of vehicles displaced to transportation calculations file and reducer
